+++ src/components/example/Example.jsx
@@ -5,7 +5,7 @@ import styles from "./Example.module.scss";
 import classNames from "classnames";
 import Card from "../card/Card.jsx";
 import Header from "../header/Header.jsx";
-
+import How from "../how/How.jsx";
 import Trending from "../Trending/Trending.jsx";
 import Footer from "../footer/Footer";
 // import User from "../user/User.jsx";
@@ -16,13 +16,14 @@ export default function Copyright() {
       <Header/> 
       <Container className={classNames(styles.container)} maxWidth="xl">
         <Paper className={classNames(styles.paper)}>
-          <Link href="/about" color="secondary">
+          <Link href="/card" color="secondary">
             Hi
           </Link>
         </Paper>
       </Container>
       <Card />
 <Trending />
+{/* <How /> */}
 <Footer />
     </div>
   );
