+++ src/components/how/Step.jsx
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import Countdown from "react-countdown";
 import millify from "millify"; 
 import Avatar from "../avatar/Avatar.jsx";
-import styles from "./Step.module.scss";
+import styles from "./How.module.scss";
 import classNames from "classnames";
 import Grid from '@mui/material/Grid';
 import Button from "@mui/material/Button";
@@ -24,86 +24,91 @@ export default function Step({
    }) {
 
 
-  
-
   return (
+ 
     <Card sx={{ display: "flex", boxShadow: "none" }}>
-    <Box
-      sx={{
-        flex: "6 0 auto",
-        display: "flex",
-        flexDirection: "column",
-        width: "30%",
-        borderRadius: "30px 0px 0px 30px",
-        backgroundColor: "#4E24F2"
-      }}
-    >
-      <CardContent>
+      <Grid container spacing={2} key={step.description}> 
+      <Grid item xs={2} >
+          <Box
+            sx={{
+              flex: "6 0 auto",
+              display: "flex",
+              flexDirection: "column",
+              width: "30%",
+              borderRadius: "30px 0px 0px 30px",
+              backgroundColor: "#4E24F2"
+            }}
+          >
+          <CardContent>
+            <Typography
+              component="div"
+              variant="h5"
+              sx={{
+                padding: "40px",
+                textAlign: "center",
+                color: "#fff",
+                fontFamily: "Montserrat",
+                fontStyle: "normal",
+                fontWeight: 700,
+                fontSize: "3rem",
+                lineHeight: "3rem"
+              }}
+            >
+              1
+            </Typography>
+          </CardContent>
+        </Box>
+      </Grid>
+
+    <Grid item xs={2}>
+      <CardContent
+        sx={{
+          flex: "12 0 auto",
+          backgroundColor: "#181828",
+          flex: "6 0 auto",
+          display: "flex",
+          flexDirection: "column",
+          width: "60%",
+          borderRadius: "0px 30px 30px 0px "
+        }}
+      >
         <Typography
           component="div"
           variant="h5"
           sx={{
-            padding: "40px",
-            textAlign: "center",
+            padding: "20px 0 0 20px",
+            textAlign: "left",
             color: "#fff",
             fontFamily: "Montserrat",
             fontStyle: "normal",
             fontWeight: 700,
-            fontSize: "3rem",
-            lineHeight: "3rem"
+            fontSize: "1.3rem",
+            lineHeight: "1.3rem"
           }}
         >
-          1
+          Live From Space
+        </Typography>
+        <Typography
+          variant="subtitle1"
+          color="text.secondary"
+          component="div"
+          sx={{
+            padding: "20px",
+            textAlign: "left",
+            color: "#fff",
+            fontFamily: "Montserrat",
+            fontStyle: "normal",
+            fontWeight: 500,
+            fontSize: "0.8rem",
+            lineHeight: "1rem"
+          }}
+        >
+          Mac Miller Live From SpaceLive From SpaceLive From SpaceLive From
+          SpaceLive From SpaceLive From SpaceLive From SpaceLive From Space
         </Typography>
       </CardContent>
-    </Box>
-
-    <CardContent
-      sx={{
-        flex: "12 0 auto",
-        backgroundColor: "#181828",
-        flex: "6 0 auto",
-        display: "flex",
-        flexDirection: "column",
-        width: "60%",
-        borderRadius: "0px 30px 30px 0px "
-      }}
-    >
-      <Typography
-        component="div"
-        variant="h5"
-        sx={{
-          padding: "20px 0 0 20px",
-          textAlign: "left",
-          color: "#fff",
-          fontFamily: "Montserrat",
-          fontStyle: "normal",
-          fontWeight: 700,
-          fontSize: "1.3rem",
-          lineHeight: "1.3rem"
-        }}
-      >
-        Live From Space
-      </Typography>
-      <Typography
-        variant="subtitle1"
-        color="text.secondary"
-        component="div"
-        sx={{
-          padding: "20px",
-          textAlign: "left",
-          color: "#fff",
-          fontFamily: "Montserrat",
-          fontStyle: "normal",
-          fontWeight: 500,
-          fontSize: "0.8rem",
-          lineHeight: "1rem"
-        }}
-      >
-        Mac Miller Live From SpaceLive From SpaceLive From SpaceLive From
-        SpaceLive From SpaceLive From SpaceLive From SpaceLive From Space
-      </Typography>
-    </CardContent>
+    </Grid>
+    </Grid>
   </Card>
 
   );
